src/db-init/model/user.ts
import mongoose from "mongoose";
import Joi from "joi";
const userSchema = new mongoose.Schema(
  {
    firstName: {
      type: String,
      trim: true,
    },
    lastName: {
      type: String,
      trim: true,
    },
    email: {
      type: String,
      trim: true,
      unique: true,
    },
    password: {
      type: String,
    },
    role: {
      type: String,
      enum: ["user", "admin"],
      default: "user",
    },
  },
  {
    timestamps: true,
  }
);

export const User = mongoose.model("user", userSchema);

export const validateUser = (user: any) => {
  const userValidation = Joi.object({
    firstName: Joi.string().required(),
    lastName: Joi.string().required(),
    password: Joi.string().required(),
    email: Joi.string().min(5).max(40).required(),
  });
  const result = userValidation.validate(user);
  return result;
};
